requirements.txt
babel==2.17.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
pillow==11.1.0
requests==2.32.3
tkcalendar==1.6.1
urllib3==2.3.0
